--- kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/ShoppingRepository.java
@@ -1,7 +1,0 @@
-package com.kodilla.good.patterns.challenges;
-
-import java.time.LocalDateTime;
-
-public interface ShoppingRepository {
-    void saveShoppingOrder(User user, Product product, LocalDateTime dateOfOrder);
-}
