--- kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/ShoppingService.java
@@ -1,7 +1,0 @@
-package com.kodilla.good.patterns.challenges;
-
-import java.time.LocalDateTime;
-
-public interface ShoppingService {
-    boolean buy(User user, Product product, LocalDateTime dateOfOrder);
-}
