+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/challenge2/informations/MailService.java
@@ -1,4 +1,6 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.challenge2.informations;
+
+import com.kodilla.good.patterns.challenges.challenge2.users.User;
 
 public class MailService implements InformationService {
