+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/challenge2/orders/ShopRequest.java
@@ -1,4 +1,7 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.challenge2.orders;
+
+import com.kodilla.good.patterns.challenges.challenge2.products.Product;
+import com.kodilla.good.patterns.challenges.challenge2.users.User;
 
 import java.time.LocalDateTime;
